requirements.txt
PyOpenGL 
PyOpenGL_accelerate
numpy
pillow
scipy
freetype-py
pyglet
typing
